improvements to animation driver
